Android: Ensure a new one time token on each login
When using the androidApiKey parameter a short-lived one time token
is generated, which can then be passed to the server where it can be
exchanged for a proper access and refresh token. Since all tokens get
cached by GoogleAuthUtil logging in a second time will cause the token
to be invalid if it was used before or some time has passed.