Hosted Domain
By submitting “hostedDomain” parameter (e.g.
hostedDomain:“mydomainname.com”) the google login will ask the user to
login using an account under that domain.
The account selection will display from the existing (in your device),
only google accounts under that domain, making the selection of the
account easier.

Therfore this doesn’t stop the user from signing in with another
account neither validates the account as member of the domain.
The user will be able to proceed to login with with any google account
if they go through the “add new account”.
It is just up to the developer to do additional check, as google
doesn’t provide an easy way to do it yet.

Note: I think that in android the webClientId is required in order the
hostedDomain to work, but in iOS it is not necessary.